Fix some code smells.

+++ src/hw/win/hw.c
@@ -1,7 +1,7 @@
 /*
  * BlueParrot XCP
  *
- * (C) 2007-2021 by Christoph Schueler <github.com/Christoph2,
+ * (C) 2007-2022 by Christoph Schueler <github.com/Christoph2,
  *                                      cpu12.gems@googlemail.com>
  *
  * All Rights Reserved
@@ -41,8 +41,8 @@
 **  Local Types.
 */
 typedef struct tagHwStateType {
-    LARGE_INTEGER StartingTime;
-    LARGE_INTEGER TicksPerSecond;
+  LARGE_INTEGER StartingTime;
+  LARGE_INTEGER TicksPerSecond;
 } HwStateType;
 
 /*
@@ -75,7 +75,7 @@ void FlsEmu_Info(void);
 void XcpDaq_Info(void);
 void XcpDaq_PrintDAQDetails();
 bool XcpDaq_QueueEmpty(void);
-bool XcpDaq_QueueDequeue(uint16_t* len, uint8_t* data);
+bool XcpDaq_QueueDequeue(uint16_t *len, uint8_t *data);
 
 void XcpHw_MainFunction(void);
 
@@ -91,98 +91,98 @@ CRITICAL_SECTION XcpHw_Locks[XCP_HW_LOCK_COUNT];
 **  Global Functions.
 */
 void XcpHw_Init(void) {
-    fflush(stdout);
-    //_setmode(_fileno(stdout), _O_WTEXT);    /* Permit Unicode output on console */
-    //_setmode(_fileno(stdout), _O_U8TEXT);    /* Permit Unicode output on console */
-    QueryPerformanceFrequency(&HwState.TicksPerSecond);
-    XcpHw_InitLocks();
+  fflush(stdout);
+  //_setmode(_fileno(stdout), _O_WTEXT);
+  //_setmode(_fileno(stdout), _O_U8TEXT);
+  QueryPerformanceFrequency(&HwState.TicksPerSecond);
+  XcpHw_InitLocks();
 }
 
 void XcpHw_Deinit(void) { XcpHw_DeinitLocks(); }
 
 uint32_t XcpHw_GetTimerCounter(void) {
-    LARGE_INTEGER Now;
-    LARGE_INTEGER Elapsed;
+  LARGE_INTEGER Now;
+  LARGE_INTEGER Elapsed;
 
-    QueryPerformanceCounter(&Now);
-    Elapsed.QuadPart = Now.QuadPart - HwState.StartingTime.QuadPart;
+  QueryPerformanceCounter(&Now);
+  Elapsed.QuadPart = Now.QuadPart - HwState.StartingTime.QuadPart;
 
 #if XCP_DAQ_TIMESTAMP_UNIT == XCP_DAQ_TIMESTAMP_UNIT_1US
-    Elapsed.QuadPart *= TIMER_PS_1US;
+  Elapsed.QuadPart *= TIMER_PS_1US;
 #elif XCP_DAQ_TIMESTAMP_UNIT == XCP_DAQ_TIMESTAMP_UNIT_10US
-    Elapsed.QuadPart *= TIMER_PS_10US;
+  Elapsed.QuadPart *= TIMER_PS_10US;
 #elif XCP_DAQ_TIMESTAMP_UNIT == XCP_DAQ_TIMESTAMP_UNIT_100US
-    Elapsed.QuadPart *= TIMER_PS_100US;
+  Elapsed.QuadPart *= TIMER_PS_100US;
 #elif XCP_DAQ_TIMESTAMP_UNIT == XCP_DAQ_TIMESTAMP_UNIT_1MS
-    Elapsed.QuadPart *= TIMER_PS_1MS;
+  Elapsed.QuadPart *= TIMER_PS_1MS;
 #elif XCP_DAQ_TIMESTAMP_UNIT == XCP_DAQ_TIMESTAMP_UNIT_10MS
-    Elapsed.QuadPart *= TIMER_PS_10MS;
+  Elapsed.QuadPart *= TIMER_PS_10MS;
 #elif XCP_DAQ_TIMESTAMP_UNIT == XCP_DAQ_TIMESTAMP_UNIT_100MS
-    Elapsed.QuadPart *= TIMER_PS_100MS;
+  Elapsed.QuadPart *= TIMER_PS_100MS;
 #else
 #error Timestamp-unit not supported.
-#endif  // XCP_DAQ_TIMESTAMP_UNIT
+#endif // XCP_DAQ_TIMESTAMP_UNIT
 
-    Elapsed.QuadPart /= HwState.TicksPerSecond.QuadPart;
+  Elapsed.QuadPart /= HwState.TicksPerSecond.QuadPart;
 
 #if XCP_DAQ_TIMESTAMP_SIZE == XCP_DAQ_TIMESTAMP_SIZE_1
-    return (uint32_t)Elapsed.QuadPart & TIMER_MASK_1;
+  return (uint32_t)Elapsed.QuadPart & TIMER_MASK_1;
 #elif XCP_DAQ_TIMESTAMP_SIZE == XCP_DAQ_TIMESTAMP_SIZE_2
     return (uint32_t)Elapsed.QuadPart) & TIMER_MASK_2;
 #elif XCP_DAQ_TIMESTAMP_SIZE == XCP_DAQ_TIMESTAMP_SIZE_4
-    return (uint32_t)Elapsed.QuadPart & TIMER_MASK_4;
+  return (uint32_t)Elapsed.QuadPart & TIMER_MASK_4;
 #else
 #error Timestamp-size not supported.
-#endif  // XCP_DAQ_TIMESTAMP_SIZE
+#endif // XCP_DAQ_TIMESTAMP_SIZE
 }
 
 static void XcpHw_InitLocks(void) {
-    uint8_t idx = UINT8(0);
+  uint8_t idx = UINT8(0);
 
-    for (idx = UINT8(0); idx < XCP_HW_LOCK_COUNT; ++idx) {
-        InitializeCriticalSection(&XcpHw_Locks[idx]);
-    }
+  for (idx = UINT8(0); idx < XCP_HW_LOCK_COUNT; ++idx) {
+    InitializeCriticalSection(&XcpHw_Locks[idx]);
+  }
 }
 
 static void XcpHw_DeinitLocks(void) {
-    uint8_t idx = UINT8(0);
+  uint8_t idx = UINT8(0);
 
-    for (idx = UINT8(0); idx < XCP_HW_LOCK_COUNT; ++idx) {
-        DeleteCriticalSection(&XcpHw_Locks[idx]);
-    }
+  for (idx = UINT8(0); idx < XCP_HW_LOCK_COUNT; ++idx) {
+    DeleteCriticalSection(&XcpHw_Locks[idx]);
+  }
 }
 
 void XcpHw_AcquireLock(uint8_t lockIdx) {
-    if (lockIdx >= XCP_HW_LOCK_COUNT) {
-        return;
-    }
-    EnterCriticalSection(&XcpHw_Locks[lockIdx]);
+  if (lockIdx >= XCP_HW_LOCK_COUNT) {
+    return;
+  }
+  EnterCriticalSection(&XcpHw_Locks[lockIdx]);
 }
 
 void XcpHw_ReleaseLock(uint8_t lockIdx) {
-    if (lockIdx >= XCP_HW_LOCK_COUNT) {
-        return;
-    }
-    LeaveCriticalSection(&XcpHw_Locks[lockIdx]);
+  if (lockIdx >= XCP_HW_LOCK_COUNT) {
+    return;
+  }
+  LeaveCriticalSection(&XcpHw_Locks[lockIdx]);
 }
 
-void XcpHw_ErrorMsg(char* const fun, int errorCode) {
-    char buffer[1024];
+void XcpHw_ErrorMsg(char *const fun, int errorCode) {
+  char buffer[1024];
 
-    FormatMessage(FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS, NULL, errorCode,
-                  MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT), buffer, 1024, NULL);
-    fprintf(stderr, "[%s] failed with: [%d] %s", fun, errorCode, buffer);
+  FormatMessage(FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS,
+                NULL, errorCode, MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT),
+                buffer, 1024, NULL);
+  fprintf(stderr, "[%s] failed with: [%d] %s", fun, errorCode, buffer);
 }
 
 void XcpHw_TransmitDtos(void) {
-    uint16_t len;
-    uint8_t* dataOut = Xcp_GetDtoOutPtr();
-    while (!XcpDaq_QueueEmpty()) {
-        XcpDaq_QueueDequeue(&len, dataOut);
-        // printf("\tDTO -- len: %d data: \t", len);
-        Xcp_SetDtoOutLen(len);
-        Xcp_SendDto();
-    }
+  uint16_t len;
+  uint8_t *dataOut = Xcp_GetDtoOutPtr();
+  while (!XcpDaq_QueueEmpty()) {
+    XcpDaq_QueueDequeue(&len, dataOut);
+    Xcp_SetDtoOutLen(len);
+    Xcp_SendDto();
+  }
 }
 
 /*
@@ -191,13 +191,13 @@ void XcpHw_TransmitDtos(void) {
  *
  */
 void XcpHw_Sleep(uint64_t usec) {
-    HANDLE timer;
-    LARGE_INTEGER ft;
+  HANDLE timer;
+  LARGE_INTEGER ft;
 
-    ZeroMemory(&ft, sizeof(LARGE_INTEGER));
-    ft.QuadPart = -(10 * usec);
-    timer = CreateWaitableTimer(NULL, TRUE, NULL);
-    SetWaitableTimer(timer, &ft, 0, NULL, NULL, 0);
-    WaitForSingleObject(timer, INFINITE);
-    CloseHandle(timer);
+  ZeroMemory(&ft, sizeof(LARGE_INTEGER));
+  ft.QuadPart = -(10 * (int64_t)usec);
+  timer = CreateWaitableTimer(NULL, TRUE, NULL);
+  SetWaitableTimer(timer, &ft, 0, NULL, NULL, 0);
+  WaitForSingleObject(timer, INFINITE);
+  CloseHandle(timer);
 }
